requirements.txt
# Azure Functions
azure-functions>=1.17.0

# Microsoft Graph SDK
msgraph-sdk>=1.0.0
azure-identity>=1.15.0

# Excel 처리
openpyxl>=3.1.2

# 날짜 파싱
dateparser>=1.2.0
python-dateutil>=2.8.2

# 환경 설정
python-dotenv>=1.0.0

# 비동기 HTTP
aiohttp>=3.9.0

# 테스트
pytest>=7.4.0
pytest-asyncio>=0.21.0
